requirements.txt
# ISA SuperApp Core Requirements
# This file contains the core dependencies required for the ISA SuperApp

aiohttp>=3.9.0
aioredis>=2.0.1
alembic>=1.12.0
anthropic>=0.7.0
asyncio-mqtt>=0.16.0
bandit>=1.7.0
black>=23.11.0

# Vector store and embeddings
chromadb>=0.4.18
click>=8.1.0
# Core dependencies
fastapi>=0.104.0
flake8>=6.0.0

# HTTP and networking
httpx>=0.25.0
isort>=5.12.0
limits>=3.7.0
PyJWT>=2.0.0
jinja2>=3.1.2
jsonschema>=4.20.0
mypy>=1.7.0
numpy>=1.24.0

# LLM and AI
openai>=1.3.0
PyMuPDF4LLM>=0.0.3
unstructured[pdf]>=0.10.0
langgraph>=0.0.40

# Data processing
pandas>=2.1.0
prometheus-client>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-mock>=3.11.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.0
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
pyyaml>=6.0.1
redis>=5.0.0
safety>=2.3.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2

# Database and storage
sqlalchemy>=2.0.0

# Logging and monitoring
structlog>=23.2.0
tenacity>=8.2.0
tiktoken>=0.5.0
torch>=2.1.0
transformers>=4.35.0
uvicorn[standard]>=0.24.0
# Graph database and analytics
neo4j>=5.17.0
graphdatascience>=1.10.0
websockets>=12.0

# RDF/SHACL Semantic Validation
rdflib>=7.0.0
pyshacl>=0.25.0

# Verifiable Credentials
didkit>=0.3.2
vc-di>=0.4.0
jwcrypto>=1.5.0
cryptography>=41.0.0
pyld>=2.0.3